+++ packages/radonis-hooks/src/hooks/useManifest.ts
@@ -7,8 +7,8 @@
  * file that was distributed with this source code.
  */
 
-import { invariant } from '@microeinhundert/radonis-shared'
-import type { Manifest, Props } from '@microeinhundert/radonis-types'
+import { invariant, isClient } from '@microeinhundert/radonis-shared'
+import type { Manifest } from '@microeinhundert/radonis-types'
 import superjson from 'superjson'
 import type { SuperJSONResult } from 'superjson/dist/types'
 
@@ -26,7 +26,7 @@ let cachedManifest: Readonly<Manifest> | undefined
  * Hook for retrieving the manifest
  */
 export function useManifest() {
-  if (cachedManifest) {
+  if (cachedManifest && isClient) {
     return cachedManifest
   }
 
@@ -34,8 +34,7 @@ export function useManifest() {
 
   invariant(manifest, 'Could not get the Radonis manifest. Make sure the server provider was configured properly')
 
-  return (cachedManifest = {
-    ...manifest,
-    props: superjson.deserialize<Props>(manifest.props as unknown as SuperJSONResult),
-  }) as Readonly<Manifest>
+  return (cachedManifest = isClient
+    ? superjson.deserialize(manifest as unknown as SuperJSONResult)
+    : manifest) as Readonly<Manifest>
 }